+++ libs/dot-components/src/lib/components/checkbox/Checkbox.stories.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
 
-import { DotRadioButton, RadioButtonProps } from './RadioButton';
+import { DotCheckbox, CheckboxProps } from './Checkbox';
 
 export default {
-  title: 'Components/Radio Button',
-  component: DotRadioButton,
+  title: 'Components/Checkbox',
+  component: DotCheckbox,
   argTypes: {
     ariaLabel: { defaultValue: 'Accessibility for the win' },
     label: { defaultValue: 'Sample Label' },
-    name: { defaultValue: 'dot-radio-button' },
+    name: { defaultValue: 'dot-checkbox' },
     onChange: {
       action: 'clicked',
     },
@@ -17,6 +17,6 @@ export default {
   },
 } as Meta;
 
-export const Default: Story<RadioButtonProps> = (args) => (
-  <DotRadioButton {...args} />
+export const Default: Story<CheckboxProps> = (args) => (
+  <DotCheckbox {...args} />
 );