+++ libs/dot-components/src/lib/components/radio/RadioButton.tsx
@@ -14,39 +14,42 @@ export type RadioSize = 'medium' | 'small';
 export type RadioLabelPlacement = 'bottom' | 'end' | 'start';
 
 export interface RadioButtonBaseProps extends CommonProps {
-  /** accessibility label */
-  ariaLabel?: string;
-  /** if the radio button is selected */
-  checked?: boolean;
+  /** id of radio button */
+  id?: string;
   /** label placement options available 'bottom' | 'end' | 'start' */
   labelPlacement?: RadioLabelPlacement;
   /** name of radio input */
   name?: string;
   /** controls the size of the radio button 'medium', 'small' */
   size?: RadioSize;
-  /** A function that should be executed when the value of the radio buttom changes */
-  onChange?: (event: ChangeEvent<HTMLInputElement>, value: string) => void;
+  /** if true user is required to select an option */
+  required?: boolean;
   /** unique value for the radio button */
   value?: string;
 }
 
 export interface RadioButtonProps extends RadioButtonBaseProps {
+  /** if the radio button is selected */
+  checked?: boolean;
   /** if true makes the radio button disabled */
   disabled?: boolean;
-  /** text displayed next to the radio buttom */
+  /** text displayed next to the radio button */
   label?: string;
+  /** A function that should be executed when the value of the radio buttom changes */
+  onChange?: (event: ChangeEvent<HTMLInputElement>, value: string) => void;
 }
 
 export function DotRadioButton({
-  ariaLabel,
   checked,
   className,
   'data-testid': dataTestId,
-  disabled = false,
+  disabled,
+  id,
   label,
   labelPlacement = 'end',
   name,
   onChange,
+  required,
   size = 'medium',
   value,
 }: RadioButtonProps) {
@@ -68,9 +71,10 @@ export function DotRadioButton({
           color="primary"
           data-testid={dataTestId}
           disabled={disabled}
-          inputProps={{ 'aria-label': ariaLabel ? ariaLabel : label }}
+          id={id}
           name={name}
           onChange={handleChange}
+          required={required}
           size={size}
         />
       }